// apps/portal-api/src/model/kanel/public/ServiceDefinition.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.ServiceDefinition */
export type ServiceDefinitionId = string & { __brand: 'public.ServiceDefinition' };

/** Represents the table public.ServiceDefinition */
export default interface ServiceDefinition {
  id: ServiceDefinitionId;

  name: string | null;

  description: string | null;

  public: boolean | null;

  identifier: string | null;
}

/** Represents the initializer for the table public.ServiceDefinition */
export interface ServiceDefinitionInitializer {
  /** Default value: gen_random_uuid() */
  id?: ServiceDefinitionId;

  name?: string | null;

  description?: string | null;

  /** Default value: true */
  public?: boolean | null;

  identifier?: string | null;
}

/** Represents the mutator for the table public.ServiceDefinition */
export interface ServiceDefinitionMutator {
  id?: ServiceDefinitionId;

  name?: string | null;

  description?: string | null;

  public?: boolean | null;

  identifier?: string | null;
}